Applied requested changes.
