+++ screen/budgets/src/main/java/com/ivy/budgets/BudgetViewModel.kt
@@ -13,8 +13,8 @@ import com.ivy.data.model.Expense
 import com.ivy.data.model.Income
 import com.ivy.data.model.Transaction
 import com.ivy.data.model.Transfer
-import com.ivy.data.model.getAccountId
-import com.ivy.data.model.getValue
+import com.ivy.data.temp.migration.getAccountId
+import com.ivy.data.temp.migration.getValue
 import com.ivy.base.ComposeViewModel
 import com.ivy.frp.sumOfSuspend
 import com.ivy.legacy.data.model.FromToTimeRange
