+++ temp/legacy-code/src/main/java/com/ivy/legacy/domain/pure/transaction/AccountValueFunctions.kt
@@ -4,8 +4,8 @@ import com.ivy.data.model.Expense
 import com.ivy.data.model.Income
 import com.ivy.data.model.Transaction
 import com.ivy.data.model.Transfer
-import com.ivy.data.model.getAccountId
-import com.ivy.data.model.getValue
+import com.ivy.data.temp.migration.getAccountId
+import com.ivy.data.temp.migration.getValue
 import java.math.BigDecimal
 import java.util.UUID
